+++ R/composite.R
@@ -2,9 +2,9 @@
 #'
 #' Create a composite image from elements of an image collection.
 #'
-#' @param x an object inheriting from `geedim.collection.MaskedCollection`, such as from `gd_search()` or `gd_collection_from_list()`
-#' @param ... [additional arguments](https://geedim.readthedocs.io/en/latest/_generated/geedim.collection.MaskedCollection.composite.html) to `geedim.collection.MaskedCollection$composite()`
-#' @return a composite `geedim.mask.MaskedImage` object
+#' @param x an object inheriting from `geedim.collection.ImageCollectionAccessor` (for geedim >= 2.0.0) or `geedim.collection.MaskedCollection` (for geedim < 2.0.0), such as from `gd_search()` or `gd_collection_from_list()`. See `\link{geedim-versions}` for more details.
+#' @param ... [additional arguments](https://geedim.readthedocs.io/en/stable/reference/api.html#geedim.collection.ImageCollectionAccessor.composite) to `geedim.collection.ImageCollectionAccessor$composite()`
+#' @return a composite `ee.image.Image` object
 #' @export
 #' @examplesIf gd_is_initialized() && !inherits(requireNamespace("terra", quietly=TRUE), 'try-error')
 #' \donttest{
@@ -23,8 +23,9 @@
 #'                resampling = "bilinear")
 #' }
 gd_composite <- function(x, ...) {
-  if (!inherits(x, 'geedim.collection.MaskedCollection')) {
-    stop("`x` should be a geedim.collection.MaskedCollection", call. = FALSE)
+  if (!inherits(x, c("geedim.collection.ImageCollectionAccessor",
+                      "geedim.collection.MaskedCollection"))) {
+    stop("`x` should be a geedim.collection.ImageCollectionAccessor or geedim.collection.MaskedCollection", call. = FALSE)
   }
   args <- list(...)
   if (!is.null(args$region)) {
@@ -34,6 +35,9 @@ gd_composite <- function(x, ...) {
   if (inherits(y, 'try-error')) {
     message(y[1])
     return(invisible(y))
+  }
+  if (inherits(y, "ee.image.Image")) {
+    return(y$gd)
   }
   y
 }