+++ R/from.R
@@ -4,7 +4,7 @@
 #'
 #' @param x character. `id` of Image, `name` of Image Collection, or a vector of Image `id` to create a new Image Collection
 #'
-#' @return `geedim.MaskedImage` or `geedim.MaskedCollection` object, or `try-error` on error
+#' @return A `geedim.image.ImageAccessor` (for geedim >= 2.0.0) or `geedim.MaskedImage` (for geedim < 2.0.0) object, or `try-error` on error. See `\link{geedim-versions}` for more details.
 #'
 #' @export
 #' @rdname from
@@ -15,12 +15,17 @@
 #' }
 gd_image_from_id <- function(x) {
   .inform_missing_module(gd, "geedim")
-  y <- try(gd$MaskedImage$from_id(x), silent = FALSE)
+  if (gd_version() >= "2.0.0") {
+    y <- try(gd$utils$ee$Image(x)$gd, silent = FALSE)
+  } else {
+    y <- try(gd$MaskedImage$from_id(x), silent = FALSE)
+  }
   if (inherits(y, 'try-error'))
     return(invisible(y))
   y
 }
 
+#' @return A `geedim.collection.ImageCollectionAccessor` (for geedim >= 2.0.0) or `geedim.MaskedCollection` (for geedim < 2.0.0) object, or `try-error` on error. See `\link{geedim-versions}` for more details.
 #' @export
 #' @rdname from
 #' @examplesIf gd_is_initialized()
@@ -41,11 +46,16 @@ gd_image_from_id <- function(x) {
 #' }
 gd_collection_from_name <- function(x) {
   .inform_missing_module(gd, "geedim")
-  y <- try(gd$MaskedCollection$from_name(x), silent = FALSE)
+  if (gd_version() >= "2.0.0") {
+    y <- try(gd$utils$ee$ImageCollection(x)$gd, silent = FALSE)
+  } else {
+    y <- try(gd$MaskedCollection$from_name(x), silent = FALSE)
+  }
   if (inherits(y, 'try-error')) return(invisible(y))
   y
 }
 
+#' @return A `geedim.collection.ImageCollectionAccessor` (for geedim >= 2.0.0) or `geedim.MaskedCollection` (for geedim < 2.0.0) object, or `try-error` on error. See `\link{geedim-versions}` for more details.
 #' @export
 #' @rdname from
 #' @examplesIf gd_is_initialized()
@@ -82,7 +92,11 @@ gd_collection_from_name <- function(x) {
 #' }
 gd_collection_from_list <- function(x) {
   .inform_missing_module(gd, "geedim")
-  y <- try(gd$MaskedCollection$from_list(x), silent = FALSE)
+  if (gd_version() >= "2.0.0"){
+    y <- try(gd$utils$ee$ImageCollection(x)$gd, silent = FALSE)
+  } else {
+    y <- try(gd$MaskedCollection$from_list(x), silent = FALSE)
+  }
   if (inherits(y, 'try-error')) return(invisible(y))
   y
 }